Test 700 (#2)
* Add test

+++ amqp_lib/amqp_wire_formatting.h
@@ -5,14 +5,16 @@
 #ifndef AMQP1_0_AMQP_WIRE_FORMATTING_H
 #define AMQP1_0_AMQP_WIRE_FORMATTING_H
 
+#include <stdlib.h>
 #include <string.h>
 #include <types.h>
+#include <stdint.h>
 
-int read_char(char *source_buffer, char *out_value);
-int read_int(char *source_buffer, int *out_value);
-int read_type(char *source_buffer, char *out_value);
+size_t read_char(char *source_buffer, char *out_value);
+size_t read_int(char *source_buffer, uint32_t *out_value);
+size_t read_type(char *source_buffer, char *out_value);
 
-int read_buffer(char *source_buffer, char *out_value, size_t len);
-int read_application_data(char *source_buffer, PMessage_t message);
+size_t read_buffer(char *source_buffer, char *out_value, size_t len);
+size_t read_application_data(char *source_buffer, PMessage_t message);
 
 #endif //AMQP1_0_AMQP_WIRE_FORMATTING_H
